Move most test high lvl api test code to test/openpgp.js, cleanup
