+++ test/crypto/index.js
@@ -1,5 +1,6 @@
 describe('Crypto', function () {
-  require('./crypto.js');
   require('./cipher');
   require('./hash');
+  require('./random.js');
+  require('./crypto.js');
 });
