+++ test/worker/index.js
@@ -1,4 +1,4 @@
 describe('Web Worker', function () {
-  require('./api.js');
+  require('./async_proxy.js');
 });
